search runtimes folder also in assembly path
