+++ FFmpeg.Native/MacOSBinaries.cs
@@ -1,39 +1,18 @@
-using System.IO;
-using System.Reflection;
-
 namespace FFmpeg.Native
 {
     public static class MacOSBinaries
     {
         public static string FindFFmpegLibrary(string name, int version)
         {
-            var assembly = typeof(MacOSBinaries).GetTypeInfo().Assembly;
-            var assemblyLocation = assembly.Location;
-            var assemblyDirectory = Path.GetDirectoryName(assemblyLocation);
+            var paths = new string[]
+            {
+                "../../runtimes/osx-x64/native/",
+                "./runtimes/osx-x64/native/",
+                "./",
+            };
 
             var fileName = $"lib{name}.{version}.dylib";
-
-            // Look for the library in the same location as this assembly. This will be the production
-            // layout (i.e. the result of dotnet build, dotnet pack,...)
-            var fullFileName = Path.Combine(assemblyDirectory, fileName);
-
-            if (File.Exists(fullFileName))
-            {
-                return fullFileName;
-            }
-
-            // Alternatively, try the "runtimes" directory. This is the layout when this assembly
-            // is loaded from a NuGet package (i.e. unit testing,...)
-            var nativeDirectory = Path.Combine(assemblyDirectory, "../../runtimes/osx-x64/native/");
-            fullFileName = Path.Combine(nativeDirectory, fileName);
-
-            if (File.Exists(fullFileName))
-            {
-                return fullFileName;
-            }
-
-            // Couldn't find the library.
-            return null;
+            return Binaries.FindLibrary(fileName, paths);
         }
     }
 }
